Fix nnpdfdb template definitions in wrapper
We have to explicitly declare BOTH the map and vector templates in
order to use them.

The attached test demonstrates the functionality working properly. It
can be run with e.g.

nosetests

The usage of ExtractMap is still awkward in Python as it is required
to instantiate a map pointer in order to pass it as output. A dict
would not be cast automatically as the object would have to be
reconstructed (expensively) at each call of the function, which is
what the input parameters are supposed to avoid. However this does not
appear to be an issue for this function. Although we could provide
such automatic casting functionality inside the wrapper, it seems
better to change the signature of ExractMap to return an std::map,
rather than modifying the input parameter in C-style fashion. Then the
resulting object can be managed with move semantics and RVOs in modern
C++.