Add script to inject changes to existing wrappers
The script works by creating a temporary git repository where it
creates the following commits:

                  |-   New template    -|
- Old template -> |                     | -> Merge
                  |-   Current sources -|

The merge then has a minimal amount of conflicts that have to be dealt
with (easy enough with meld).

Currently the swig.template file has to be edited in the middle of the
process, since we need to know how the wrappers look like when
generated with the old template. The script will prompt the user to
edit the file.