Developer preview of the data theory comparison
This uses partly the machinery of the future validphys2. Several bits
of the functionality have been backported from SMPDF, but with a more
general (and also more limited at the moment interface).

At the moment we can parse a yaml file, such as that in the example
and produce the corresponding plot comparing data and theory.

At the moment no easy setup or configuration options exist and the
user is responsible for managing the dependencies (this will change
in the future). In particular they need to setup the libnnpdf wrappers
and reportengine.

The interface (both API and UI) will change in the future
substantially and it is advised to not rely on them for the moment.