Implement PDF and Results infrastructure
Implement a PDF class and a set of Stats classes, closely following the
SMPDF implementations.

The PDF object can query the info properties, and load the NNPDF
object with the `load` method (but with the correct error types).

The result infrastructure is a little more cumbersome (and most
probably not final) because it has to combine the NNPDF calculated
with the python calculated results.

We implement a set `stats` objects that adequately represents the
errors for hessian and mc sets, and pass things like the chi2 data
trough it.

NNPDF dropping silently replica 0 for MC sets seems like a bad idea as
can be error prone. Have to think of a better way...