Add support for cuts and add refactor Datasets
We move all the dataset related logic (except the actual NNPDF object)
to a single class so that we avoid passing argument around everywhere.

This helps adding support for cuts, for which we require to locate the
fits:

To locate the fits, we add another executable option, but calculate
the likely nnpdfbuild folder from the datapath we already had.

For the moment parse_fit just returns the fit path, but should do more
eventually.

With this we load the cuts from the specified fit if the use_cuts key
has been set to true. If it is false, a fit is not required (this
requires the latest version of reportbuilder to work).