Use reportengine loop in validphys
We split validphys in logical components and use the reportengine
driver to parse the configuration file and generate the plots.

In particular, the targets are loaded from the actions_ tag, and are
iterated following the rules of reportengine (a dict enters the stack,
a list does that for each element of the list).