Add components to parse plot options
The new library consists of a PlotInfo specification and parser,
together with facilities to compute the table which would be used for
filtering.

The plotting options are passed trough one or more YAML files per
dataset (in the data folder).

This file contains:

 - A set of extra_labels to be interpreted as extra specifications for
   each datapoint (describing the process or acting as an extra
   kinematic).
 - A transformation function for the original kinematics k1, k2 and
   k3, which is defined in the `plotoptions.transforms` module.
 - The label which will be acting as x.
 - The labels used to group different figures.
 - The labels used to group different lines in a figure.

Everywhere where a label is required, a function defined in the
`plotoptions.labelers` module can be used as well. It takes as
arguments k1, k2 and k3, as well as any extra_labels defined in the
YAML file.

An example of configuration file can be found in the tests.