Make PDF and DasetSpec comparable and hashable
This is going to be required in several places, like for example
dealing with a "normalize to" specification.

The PDFs and datasets are considered equal if their defining
parameters (but not ids) are equal, in a way that load() returns an
identical object.  This allows us to have a cache in load.

We try to return tuples
instead of lists everywhere where they are not expected to change.
This is not required for the comparison, but it is for the hashing.
Additionally, it makes the intend more clear.

Still not sure what to do with the cuts. Converting a potentially
large array into a tuple. One option would be to link
the filename.