Implement normalize_to for fancy plotting
This implementation is probably not optimal. This is because:

 - It relies on the rather demanding check interface to arbitrary
   transform the input.

 - It is not very clear what the interface of the plot_fancy function
   should be, e.g. if one wishes to compare multiple theories on the
   same plot. Also supporting both 'pdf' and 'pdfs' should be done
   automatically somehow, without having to branch for this case.

 - Supporting normalize_to on the same function as the non normalized
   case makes it much more complicated to follow as the control flow
   branches multiple times.

 - It doesn't take advantage of the rest of the framework and
   reimplements thing and adds special cases (e.g. the key 'data') on
   its own.

However, while waiting for a better spec to emerge, this is what it
is: The normalization always refers to one of the computed Result. It
can be specified by the special key 'data', which refers to the index
0, the DataResult as per the current convention, or the PDF id (which
must exist in the namespace). Alternatively, one can just pass the
corresponding index of the list of results (e.g. normalize_to: 0 means
the data result, 1 the result for the first PDF and so on).