Implement data members for DataSet and its parents
Implement copy assignment and move constructors (which come for cheap)
for CommonData, FKSet and DataSet that work in the way other cpp
clients (e.g. SWIG) expect them to:

 - Copy assignment copies the assigned instance into the assignee and
   cleans it.

 - Move constructor depletes the resources of the argument and assigns
   them to th new instance (but without expensive copies).

The implementation is largely based on:

http://stackoverflow.com/questions/3279543/what-is-the-copy-and-swap-idiom

It turns out that by implementing the copy assignment operator in terms of
the swap function, the move assignment comes for free always
(although it is not the most efficient implementation, it is still
quite a lot better than copying).

The move constructor can also be implemented in terms of swap. This is
done for FKSet and CommonData, but not for DataSet (because it's more
complicated).

Some changes that are needed are:

 - Remove const for private members because we need to be able to
   assign to them.

 - Some checks in the destructor to make sure they work properly for
   0 data.

This commit deliberately does not use the new functions anywhere in
the code, even though they can be used to optimize e.g. the experiment
class.

The correctness of the new functions is tested for with a program in
the "tests" repository.