Add lru_cache to the load() methods
This avoid loading over an over the same object from the code. It
requires making sure that each of the relevant specs in hashable (so
experiments save datasets as tuples and not as lists and so on).

It turns out, this works perfectly well for execute parallel, as we
only need to load the resources once per process.

The default size of 128 seems about right.