Add methods to export tables
Add actions to export:

 - Data central values and thory predictions.
 - Data covmat and inverse covmat.

These are accessible as validphys actions. They use a helper function
that creates a fancier index with all the relevant info.

Right now the code for indenxin and predictions is not very
performant, and it should remove the avoidable python loops in favour
of something more efficient (like is done by setting the values
directly in the covmat methods).