Implement vector_thpredictions in wrapper
This is done like vector_dataset, with the difference that there is an
import cycle to break. At this point it is a mystery which is the
right combination of:

 - %include
 - #include
 - %import

That does the trick, so it is deduced by a time consuming trial and
error process.

As with the other classes, ignore the new ThPredictions data methods,
as they are useless for the wrapper.