Catch better StopIteration exceptions
If these are thrown inside a generator, the generator exits silently
at that point, and the error becomes invisible. To remedy this,
there is

from __future__ import generator_stop

so insert it in all the files containing yield.

This is particularly annoying when matching the kinematic labels in
the plotting mechanism (which is itself fragile and relies on the
correct settings been written in the commondata.cc file). Thus, give
a more informative message there.