Fix not found exception reporting
A FileNotFound exception is triggered by the available_fits function,
which in turn is called by the exception raised in check_fit. So it
ended up raising the FileNotFound exception instead of the ConfigError
exception resulting in a confusing traceback instead of a nice error
message.

We simply return an empty list in the case we get an IOError, which
does seem like a reasonable behaviour.