Implement automatic binning and pass by array
We pass arrays instead of single numbers to the user defined
functions. This is likely to be faster, but more importantly it allows
for automatic binning.

The bins() function will do the right thing if the data has identical
values in some label, thus we don't have to implement extra_labels
just for the binning anymore.