Add fitdata capabilities
We copy some parts of the postfit2 module to read and analyze
replicas, in particular to read fitinfo files.

The plan is to make potfit2 depend on this.
We add a validphys style check to try to make sure that postfit ran
(and require the PDF to be installed for good measure).

We add a toymodel-like training-validation plot (although without the
marking capabilities for now).