Remove derived build files
I am removing those because they are essentially not portable from one
system to another. I keep getting emails from people running the
configure file with an incorrect configurations, and they make the git
logs very messy.

This way we add an extra dependency to build the library, but it was
there for all practical purposes. You need a working build system to
build the library!

Furthermore the compiled packages reduce the pressure to have
these for end users.

Conflicts:
	.gitignore
	configure