Implement global verbosity control
The verbosity was too granular and inconsistent in that it was
not applied to all the prints.

Instead, implement a global logger object that returns either stdout
or a dumb ostream, depending on the value of a Verbosity variable. It
could be easily extended to e.g. return a file stream.

Also, implement getters and setters for that global variable. Main
reason is that SWIG didn't like the bare global flag, nut this way
allows to do nontrivial stuff with the setter in the future.

For the moment, do not touch some prints that could look like serious
warnings, although in the future could be good to have also a warning
interface.

LHAPDF needs a separate verbosity control which is implemented in
SWIG (nothing to be done for c++). This is needed because the lhapdf
python wrapper doesn't talk to the C++ side.