Support positivity prediction
This entails a new class of Result where the behaviour is determined
by the stats type (of the pdf).  This looks like the way it should
have been at the beginning except for the duplication with the NNPDF
code. It is not clear it makes sense to maintain both approaches.