Add meson build file
The library builds ~7 times faster on my computer and I can actually
read the generating files.

As a bonus, it generates pkgconfig files which come handy for linking
the library (eventually no need for nnpdf-config).

The whole build system could be easily deployed with the installer
tools eventually.

Wrapper support is missing, as meson currently does nothing about
python (and just calling seyup.py misses the point of just rebuilding
modified files).

For the moment this is experimental and for personal use.