Use the Cholesky decomposition to calculate chi²
In addition factor out the two lines that actually do the computation
so they can be easily used for any input.

The inverse covmat that can be exported is probably not very useful
now, so it doesn't make sense to search for the LAPACK call that does
triangular inversion (can always be done in the future).

I don't link the chisquared module in libnnpdf, because it makes
little sense to implement backsubstitution by hand and even less to
write wrappers for it.