Add plot_pdfreplicas
Add an action to plot pdf replicas. The plotting function has
automated axis scaling, support for linear and log scales, and
support for results normalized to a given PDF.

All arguments have checks incorporated in the validphys app.

The function is written so that alternative representations of the
same info (error bands) are easy.

We modify pdfgrids so as to have information on the distribution of
the x points and use that to set the default plotting scale.

We can thread grids over pdfs with the new reportengine collect
functionality.

We also add a label property to the pdf to be easily used for
plotting purposes.