Add PDF plots
Supports mc and hessian sets, normalization and various scales like
the replica plot.

There is a large amount of code shared between this and the replica
plot, at the cost of a rather ugly factory interface. There is also
the issue that an inner function can only be pickled (which we need
for execute_parallel) if it wraps some named function in the module.

Still, this factory can serve as example for how to deal with
plot_fancy.

These are largely inherited from SMPDF, but improved: The plotutils
functionality for the hatches and the legend handles is copied
verbatim. The axis limits algorithm is far better, and also hatches
work in the PDF format now.