Calculate chi² using collect
This enables results to be computed in parallel when calculating
chi²s.

There is a certain ugliness streaming from the necessity to define
the same function several times for datasets and experiments, but for
the moment this is easier than reportengine gaining yet another
gigantic framework for multidispatching. In any case some more
experience with `collect` is needed before deciding.