Implement remote loading
Add the capability to download missing fits, pdfs and theories
automatically. This is enabled by default and controlled with the
--net/--no-net flags.

Now when a e.g. a pdf is not found, it will be downloaded and
installed before proceeding to analyze the rest of the graph.

The nnpdf resources use "index files" (an idea borrowed from conda)
that we now have to regenerate automatically, and extend with e.g.
metadata on what's in the theories that are updated at random now.
For LHAPDF we use the local index file (should we call lhapdf update)
but not the command line because it doesn't give any useful
information on failures (lhapdf install patata has exit status 0).

We use a FallbackLoader class that uses some python magic to try to
download resources every time we check for them.

This can improve in that it breaks the premise of not doing any work
before checking correctness as much as possible. Ideally downloads
would be asynchronous and synchronized with node building. This
requires some usable http library for curio (instead we use requests
for now which is nice and well established) and a reorganization of
ResourceBuilder so that the inputs are nodes that are built like
anything else (for which maybe the 3.6 async generators could be handy).