Add --upload option
This will upload the report via ssh/rsync to an unique url of the
Milan server.

It will perform a (expensive) check of the connection (and the rsync
executable) at the beginning and exit with error if it doesn't work.

Possible future improvements include generating an index page for all
the uploaded reports based on some metadata (like e.g. the author)
streaming from nnprofile.

When finalized, the URL will be shown to the user.