Add CutSpec
This improves on the ridiculous frozencuts that we had in the
DataSet. However the main motivation is that allows to separate the
cuts from the dataset. This will allow doing convolutions without the
cuts (good for performance since copying the dataset structure is
very expensive) and attaching the cuts to other objects such as
CommonData (good for e.g. kinlimits).