Add vp-upload command
This is useful to quickly upload an existing output folder.

Why a new command?

Mostly, because as it is, validphys requires a config_yml argument,
and is not clear to me what the semantics of a --just-upload flag
would be.

Why is it not called validphys-upload?

Because I don't want to type an extra space every time I tab-complete
validphys. The downside is that the new command is more difficult to
discover, but that's what the docs are for.