Add SMPDF plots
This adds the machinery for computing correlations ala SMPDF (Monte
Carlo sets only for now) in an optimized way without Python loops.

As a further improvement from SMPDF, the color bar now shows the
value of the x axis for each bin.

Some minor improvements could be added to the plot (e.g. color bar at
the bottom).

A more important problem is that the plots are painfully slow. This
is likely due to some performance regression in matplotlib.

Some common plotting functionality has been refactored, though the
amount of code reuse between plots is still low (not sure what can be
done about it).