Fix moving of invalid replicas
Move the invalid replicas to a .invalid folder.

This may crash running jobs (which is why I didn't do it before), but
it's better than having postfit randomly moving one replica folder
inside the other. The solution is to change nnfit to write in e.g
replica_XX.running, but that's for later.