Fix race condition in server script
The action of the index script would create new inotify events, that
in turn trigger the script again, resulting in a high resource usage.
The fix is to continuously watch for the first event and only create
a new watch after the script is done.

There are probably nicer ways than creating a new watch descriptor
every time, but this is safe and clear.