Support .tar.gz files only
Detecting arbitrary compression formats depends on heuristics that we
don't really need to rely upon. This is especially problematic if we
rely on the fallback mode to use the plain files, since the fallback
can be pretty long, and the decompression can fail at different
stages (apparently some fktables are valid compressed archives in some
format while others are not).