Support producing CommonData and Cuts directly
Loading the entire dataset with the fktables is extremely expensive
when not needed. Some actions like the kinlimits are much fasted
if only the CommonData part is load()'ed.

The proposed solution is not very good in that the name attribute is
dangling awkwardly. Maybe it would be better that DataSetSpecs gets
it from the cd directly. In any case this is minor if everything is
going to be provided by the Loader as it should.