Create lumigrid2d provider
This computes the luminosity for the allowed values given sqrt(s) and
does not do any statistics.

Implement for one channel and use collect() to iterate over channels.

Merge this with the old evaluate_luminosity_grid (which was providing
two lines worth of abstraction).

No checks are needed because its all handled by the type checks and
the config checks (although it would be useful to have a check for
positive values).