Plot lumi uncertainties for a singe channel
Old behaviour can be obtained with

{@lumi_channels plot_lumi2d_uncertainty@}

There are arguments for doing it both ways (and eventually we could
have a collect function that does both) but this looks slightly
neater and reduces the amount of explicit control flow. Also,
compared to the PDF plots, there is less common setup code and less
performance adantage in doing it all together.