Add pdf production rules
This is an ugly hack. Basically it finds a pdf defined in a base
mapping and appends it to other pdfs (to add insult to the injury it
comes in two versions). I'll need to find a better way (like an
`append_from_` key) but that will have to wait.