Add fitinputcontext production rule
This is useful for analyzing a fit that hasn't completed a PDF (e.g.
to check that the cuts are sane).

It should be possible to move these things away from Config to
somewhere where they don't take brain space. This rule is probably
not generally useful in config files, but good to use with collect.