Change PLOTTING to load type files
Now the loader will search for one data specific file and one process
type file. Anything in the data file overrides anything in the process
file (might have to think how e.g. normalize should work, but this is
largely an "academic issue" right now). For this to work, we introduce
a way of peeking the process type that is faster than processing all
the commondata at compile time.

Also, add process_description, which is not used anywhere at the
moment.

Now the interface of get_infos doesn't make much sense since we only
have one PlotInfo object.

This is the first commit that requires Python 3.6!

Fixes #10.