Change the names of the highlight keys
Better do this ASAP before people start depending on these names. The
rationale is that there should be an interface where a given scatter
plots can have many 'marks' and only one highlight. Then what is
marked and what is highlighted is computed on an action by action
basis, based on the user input. Then it makes sense to have a single
"highlight_label" and multiple ways to obtain highlights. In this
case we are giving in the dataset names, so we call it
highlight_datasets.

Needless to day we are not there yet, and the functionality in
plot_xq2 is fairly ad hoc, but it provides a test case for what
a generalized API should handle.