More work on as fitting
This adds an action that plots the replicas by id and performs one as
determination per replica id. The central value and uncertainty are
displayed in the title and the colorbar shows the preferred as value
of each replica.

We also relax the requirements for fits_pdfs so that one doesn't need
to download the theory.