Add bar plots that compare chi² for fits
We also refactor the fits_chi2_table function so that it shows both
the total chi² per experiment as well as the per-dataset one, and is
based on two separate tables, which are providers for these functions.

The new functions are:

 - plot_fits_datasets_chi2
 - plot_fits_experiments_chi2

and only require a list of fits.

It would also be useful to have versions that  show the percentage
contribution to the total chi^2, to show what is the fit having
trouble with.

This API is convenient but not very flexible. Next is to allow a free
form  ``data_specs`` key that implements this function for some
arbitrary combination of experiments, theories and PDFs.