Add option to enter experiment_input
experiment_input is different from experiment in that it doesn't
require any external key to be parsed. Then it results in an
Experiment (that takes the theorid and cuts) when using explicitly
the production rule "experiments_from_input". That this is
not implicit (produce_experiment()) because currently the interaction
between production rules and parsing rules of the same thing is not
particularly nice.
Possibly a better approach would be to translate all instances of
"experiment" to "experiment_input" in the config file, before it even
gets processed.

It is not clear how this should look like ideally.