Add experimental dataspecs API
Having providers that iterate for fits is convenient but not very
flexible. We need a way to specify arbitrary values of experiments,
theory and pdf. Then the various providers just loop over those user
defined lists.

This needs a little bit of work to be polished. Right now a major
annoyance is that it doesn't compose well with from_. For example, it
would be nice to be able to write:

myspec:
 - fit: A
 - fit: B

override:
   pdf:
     from_: fit

dataspecs:
    allfrom_: myspec::override

as opposed to

dataspecs:
 - fit: A
   pdf:
      from_: fit

 - fit: B
   pdf:
      from_: fit