Add compile time rule for matched data
This is a crazy piece of logic. It goes as follows:

The user defines something called "dataspecs", which can be any list
of dictionaries. For example:

dataspecs:
   - theoryid: 53
   - theoryid: 54

Then there is the "mathched_datasets_from_dataspecs" production rule
that produces "experiments" within each dataspec (so in the example
above it would pick different fktables), then it matches them by the
name of the experiment and the dataset and returns a  mappings
for each of the results in the intersection which contains the
following keys:

 - dataset_name
 - experiment_name
 - dataspecs

The inner key "dataspecs" (not yet sure it has to have the same
name) contains a list with mapping per dataspec, with the complete contents
of the dataspec as defined (e.g. theoryid:53 for the first one above)
and additionally:

 - dataset: The dataset as resolved by the dataspec, with the matched
   dataset_name.