Reorganize the commondata metadata handling
Get basic properties by peeking at the file, without having to load
it (in particular process it with streams). In total this is slightly
more time, as we potentially open the file twice, but since it's
worth it since we get the information at compile time and the file is
going to be cached by the OS the second time.

The functionality is moved to validphys.core (from plotoptions) so
there are less weird interdependencies. Eventually we should remove
the references to the kinlabels, which are only there because there
is no one to one mapping between library process type and the
validphys process type.