Revert back to vector<Experiment*>
SWIG requires implementing copy assignment for Experiment in order to
use it for a vector, at which point its better to just rewrite it. But
we said that's too much work.

Also add a way to access the vector of datasets rather than the
individual elements, mostly to make looping simpler.