Rework plot_fancy and add a datasepec interface
This is still not the "final" version and requires a few changes in
libnnpdf to be really nice (in particular a class that doesn't know
about theory but knows about covariance matrices).

We add a "private" implementation, that serves both for the single
dataset-many-pdfs case and for the more general dataspecs case.

We treat the cuts explicitly in the plot_fancy implementation,  so
that we can have data with different cuts. We take explicitly
a commondata object and compute the whole kinematc table without cuts.

For the dataspecs, we allow to customize the labels by defining
a "speclabel" parameter for each spec. We need an extra bit of logic
to discard empty figures (where all the points have been cut). This
is not particularly comely. Should look into making it better.