Rearrange iteration in GenCovMat
Loop over systematics first and correlated points second only if
needed.

We trade a lot less string comparisons with some more multiplications
(good) and more indirect writes to the covmat (not so good). It turns
out to be a lot better now, but that might change both ways if we make the
datastructures more compact and use enum for comparisons.

Also obviously it depends on the relationship between the total
number of systematics and the uncorrelated ones. The more the
uncorrelated the better this approach is.

The net effect is that the pseudodata benchmark goes from 10s per
replica to 7, with the DataSet GenCovMat method dominating the waste
of time now.