Work on alpha_s fitting tools
At the moment these are based on running the results from file,
though this shouldn't be needed all that much if we fix the issues
with the covmats.

The produce_fits_matched_pseudorreplicas_chi2_output_by_experiment
basically converts the data file to the same interface we had
previously as so we can use the same functions, which in turn have
been refactored into smaller more sensible providers.