Build the wrapper in one step and from meson
Compile a single wrapper file from one source instead of per
header.

Now the compilations is much faster. While it's not strictly required
I prefer to have the wrapper definitions in one file, instead of in
many, because there is no such thing as separated logical units in
SWIG and instead everything depends on everything else.

Also add "smart" meson definition allowing to rebuild the wrapper
exactly when one file has changed and avoiding the nefarious
setuptools setup. At this point meson looks nice, although I had to
work around several issues. The main problem was getting it to copy
a bunch of files to a given folder. One problem remaining is that one
needs to point PKG_CONFIG_PATH to the conda location for things to
work as expected (setuptools gets the relevant python flags, but meson
doesn't for some reason).

Also `pip uninstall NNPDF` doesn't work, but it was broken anyway.
Instead `ninja uninstall` does work.

CMake still works with that setup (but also compiling only one
wrapper).

One change is that now all the definitions come from one module, so
there is no such thing as

    from NNPDF.thpredictions import ThPredictions

Because this would break validphys and possible other uses of the
wrappers, we add backwards compatibility modules that print a warning
and then forward to the main module (and are all symlinks to the same
file). This should keep things working while they are updated.

Fixes #12.