Add regression tests
The tests use the validphys actions to save files to disk and then
check that the content hasn't changed.

Compared to C++ tests, they have the advantage that are higher level
(so presumably more easily maintained), easy to write and more
powerful, and the disadvantages that they introduce more layers, it is
more complicated to debug things like memory issues and require
a synced version of validphys. These particular tests check that the
numerical results are maintained between refactorings.

They are meant to first be run
on a branch that is supposed to work, like:

    cd tests/vp
    py.test

The tests will fail and generate a tables with the data. On a second
run of `py.test` the results will be compared with the saved ones.

Writing regression tests is very easy thanks to some magic with
decorators, but complicated by the fact that pandas doesn't roundtrip
CSV particularly well. Perhaps we should use a different format.

Th first approach was trying to make the interface based on validphys
input cards (so the tests are even simpler to write), but the
functionality of pytest is hard to integrate in the reportengine loop.
Perhaps in the future. Anyway the pytest fixtures works very similarly
to the reportengine reportengine providers, so code that is good for
one should be good for the other.

Note that this requires validphys master.