Add table loader module
This is to process the output of the table providers back into pandas
dataframes. I'm not sure this is a good appoach in that CSV files do
not roundtrip well at all.

In any case, it is good to keep this separated from validphys.config
so that, similarly to Loader it can more easily be used in scrpits.