Implement pseudodata chi^2 for datasets
This now can work "reasonably" fast (as in 95% of the cpu time is
wasted in utter garbage), and should be "sufficient" for computing
the per-dataset chi².

The companion function in results.py is incredibly ugly and full of
holes, so we will have to see how to restructure it better. But at
least it seems to work.

The format of the output table has changed so the readers will have
to be adapted as well.