Add more table loading pipelines
I am not sure at all I like this interface, but these are needed now,
in order to work with the alpha_s fits.

The idea is that the output files have to be "enabled" with a {@with@}
block, and then the tables require to pass them the right fit names,
which is useful both for filtering, and to check that we haven't done
anything completely crazy.

Also rename the parse_* functions in tableloaders to load_* to make
them more descriptive and more consistent with loader.