First implementation of wiki-upload
This supports single files and asks about the meta.  prompt_toolkit is
a new dependency. We can make better use of it than this, since it
supports completions and multiline input. However it tuns out it is
mostly useless for a simply confirm message and one has to undo most
of what it does by default.

There are various subtleties with symlinks and permissions that seem
mostly working now.

I am still unsure about many details.

Fixes #40.