Add fast methods to access xfxQ in Python
I am less than sure I want to introduce a new toolchain now that the
wrapper was looking almost nice, but I believe we are going to need
these quite a bit in the long run.

Things like scipy (since very recently) and numba support calling
"CFFI" functions directly as C code and thus avoid the overhead of
calling a Python function every time. This is important for things
like integrals where a function is called many times in a loop.

For something with a reckless disregard for performance, such as
LHAPDF.xfxQ, this overhead is 100%-200%. For a better optimized
function, it can be more.

In SWIG we have something like grid_values, which does the loop
itself in C, but that doesn't work with all the APIs.

While we could do these all the way in C++, I was put off by the fact
that the APIs are not super nice (like IntegratePDF taking a mandatory
"gsl_integration_workspace*" for some reason). Besides this opens the
door to interact with e.g. numba.

Instead, introduce another wrapper technology, based on C, for the
moment, essentially to wrap xfxQ.

Some examples of places where this can be used are:

 - arclength evaluations.
 - sum rules
 - luminosities

It works similarly to SWIG but with less magic (for the good and the
bad) and only with C headers (which is why less magic is required). It
generates a C/C++ source with the binding boilerplate, from an input
source and an explicit list of declarations to be exported.

To make this interact with SWIG objects, one has to convert the
pointers to integers of the appropriate type and then to pointers CFFI
pointers again. There are a few one liners with integrands in the
format scipy likes them. The whole thing is to be used like this:

https://github.com/NNPDF/validphys2/commit/2ed324ccb6d4411c4419226b98e78aae399d1ce7

or as in the test file.

The net effect is that the Python overhead is completely removed and
integrals work at "C speed". See for example for the sum rules:

http://pcteserver.mi.infn.it/~nnpdf/validphys-reports/ooD25rYbRkOyW7ny6CgyWw==/sumrulesperf.svg

The tests pass building with both meson and CMake.