Add ability to dowload validphys outputs
This is useful for e.g. the as fits where the chi2 are produced in
one place and consumed in another. Now the paths can be specified
relative to the validphys reports root, and will be downloaded if not
found in some "validphys cache directory".

The greatest advantage is that this way the derived results are
trivially reproducible.

More work could be done to polish the interface, but this seems good
as is.

The versions of reportengine and libnnpdf have been bumped since we
require some new functionality.