Make access to the pathlib define volatile
The problem is that some compilers (clang on Mac) want to be smart
and optimize the length of the string.

But then, the string changes on the binary itself when conda does its
relocation mess.  Specifically conda builds the files using a very
long $PREFIX, which then gets replaced by the user prefix at install
time, which will have a shorter length. So if the string mindlessly
remembers its length, the user ends up with a lot of zeros in the
path.

I have no idea if this will work.