Add  loading of the minumum table
We add a way to compute the minimum of N sets of chi2 and store it to
a table.

These are done quickly and without thinking too much. In particular,
there is the problem that currently I don't know how to compute the
minimum and the rest in one go, so first we write a facility to
export the table, and we load it in a subsequent run. At least thanks
to the validphys files, the table can be computed once and stored in
the server. Future versions of validphys and reportengine should bring
improvements to this area.

Also, one has to do various gymnastics to compute
the total.