Use nnprofile URLs
In particular, this includes CERN URLs. Adding support for this is
straight forward, except that the Apache server at CERN is improperly
configured, and causes the code to think it has to uncompress the
archives twice. This requires a rather cumbersome workaround.

Right now, the updating of the indexes in the CERN repo has to be done
manually.

Fixes #37.