Deal with different header grid shapes in lhio
This allows to compute replica 0 also with sets where the grids of the
members are not the same.

This required re-enabling the reading based on the LHAPDF library
(which is used directly, instead of having to rely on the mess that
libnnpdf does with the indexes).

Took the opportunity to modernize the code a bit.