Remove explicit CXX setting from Makefile
According to the documentation:

https://www.gnu.org/software/make/manual/html_node/Implicit-Variables.html

`CXX` is an implicit variable that does not need to be overridden to
work. Not setting it allows users to set their own compiler, and in
particular allows to remove a special case from here:

http://pcteserver.mi.infn.it/~nnpdf/validphys-docs/guide.html#development-installs

where in order to function properly, users need to be changing ugly
names.