Fix bug in the computation of number of points
We were grouping by number of points and then summing the items. When
the points in two consecutive datasets are the same, they will be
grouped together and counted only once.

The correct behaviour is to group by dataset and get the unique number
of points.

The more correct behaviour would have been using parket or something
that has composed data.