Break paramfits into submodules
Fixes #51.

This breaks:

 - The Paramfits specific config away from the main config.
 - The plots away from the rest of the providers.

Noe paramfits.py is split into paramfits.dataops and paramfits.plots.
One may still split dataops in more parts (e.g. things that do
computation and things that move data around) but this is still some
improvement as it reduced the number of lines of all the files to
a manageable size, with more or less a logical split.

All the alpha_s runcards in the paper repository seem to be
unmodified.